BeaconFrame parsing itself
